Update Python Derivative Calc. Text File
Here is my code for a general derivative calculator; for the most part, the logic works. I need to implement some of the more obscure transcendental functions (and some inverse functions), but the main part of the program works. It's in a .txt document because I use Jupyter notebooks to run my code in a command-line style of editing. That's probably a pleb move, but I'm still relatively new to Python, so that's just how it's gonna be.